+++ packages/polar/src/pie/predicates.ts
@@ -12,7 +12,7 @@ export const isPieProcessedData = (data: Array<unknown>): data is Array<PieProce
     const result = data.map((item: unknown) => {
         if (typeof item !== 'object' || item === null) return false
         if (!('id' in item) || !('index' in item)) return false
-        return 'startAngle' in item && 'endAngle' in item && 'value' in item && 'proportion' in item
+        return 'startAngle' in item && 'endAngle' in item && 'data' in item && 'proportion' in item
     })
     return result.length > 0 && result.every(Boolean)
 }